Rearrange lib.rs and move pretty_assertions module into it's own file
